Create Grade the Student
You are given a character ch which represents a grade of a student in a course. The valid grades are A, B, C, and F. Your task is to print a message based on the grade using a switch statement.

If the grade is A, print Excellent!.

If the grade is B, print Well done!.

If the grade is C, print You passed!.

If the grade is F, print Better luck next time!.

If the grade is not one of the valid options, print Invalid grade.

Input Format

Input contains character ch.

Constraints

A <= ch <= Z
Output Format

Print the desired output.

Sample Input 0

B
Sample Output 0

Well done!
Sample Input 1

A
Sample Output 1

Excellent!