Create Switch Calculator
You are given integer inputs N , a and b. perform operations on a and b for different value of N like :-

If value of N is 10 -> a + b;

If value of N is 20 -> a - b;

If value of N is 30 -> a * b;

If value of N is 40 -> a % b;

If value of N is 50 -> a / b;

else print Enter a valid number.

Input Format

first line of input contains integer N.

secind line of input contains integer a.

third line of input contains intager b.

Constraints

0 <= N <= 50

0 <= b <= a <= 10^4
Output Format

Print the desired output.