Create Print the Day
Write a Java program that prompts the user to enter a number between 1 and 7 (inclusive) that represents a day of the week. The program should then use a switch statement to output the corresponding name of the day of the week.

for 1 output is Monday , for 2 output is Tuesday and so on.

Input Format

Take input N form the user.

Constraints

1 <= N <= 7
Output Format

Print the desired output.