Create Print the Month
Write a program that prompts the user to enter a month number (1-12) and uses a switch statement to display the name of the corresponding month. If the user enters an invalid number, the program should print Invalid Number.

for 1 print January , for 2 print Feburary and so on.

Input Format

Take integer N as an input.

Constraints

0 <= N <= 100
Output Format

Print the desired output.